Fix incorrect type of sec_per_clus
In the definition of `fat32_t`, `sec_per_clus` was defined as
`uint32_t`.
According to the FAT32 specification, it only occupies 1 byte and
should be defined as `uint8_t`.
While this change has little practical impact, it avoids confusion
and corrects the type.

+++ include/type.h
@@ -46,7 +46,7 @@ typedef struct {
     uint32_t first_fat_sec;
     uint32_t first_data_sec;
     uint32_t byte_per_sec;
-    uint32_t sec_per_clus;
+    uint8_t sec_per_clus;
     uint32_t root_clus;
 
     uint32_t fs_info;
